moved oats_subtypes into oats types
