+++ bin/migrate-oats-data/common/__init__.py
@@ -1,3 +1,3 @@
-from .oats_application_enum import *
+from .oats_application_code_values import *
 from .alcs_application_enum import *
 from .etl_logger import log_start, log_end
