--- bin/migrate-oats-data/common/oats_application_enum.py
@@ -1,29 +1,0 @@
-from enum import Enum
-
-
-class ALRChangeCode(Enum):
-    TUR = "TUR"  # Transport, Utility, and Recreation
-    INC = "INC"  # Inclusion
-    EXC = "EXC"  # Exclusion
-    SDV = "SDV"  # Subdivision
-    NFU = "NFU"  # Non Farm Use
-    SCH = "SCH"  # Extraction and Fill
-    EXT = "EXT"  # Extraction
-    FILL = "FILL"  # Fill
-    SRW = "SRW"  # Notification of Statutory Right of Way
-    CSC = "CSC"  # Conservation Covenant
-    NAR = "NAR"  # Non-Adhering Residential Use
-
-
-class OatsCapabilitySourceCode(Enum):
-    BCLI = "BCLI"
-    CLI = "CLI"
-    ONSI = "ONSI"
-
-
-class OatsAgriCapabilityCodes(Enum):
-    P = "P"
-    PD = "PD"
-    MIX = "MIX"
-    S = "S"
-    U = "U"
